+++ sync.sh
@@ -7,8 +7,10 @@ cp ${HOME}/.xinitrc ${CDIR}/
 cp ${HOME}/.xbindkeysrc ${CDIR}/
 while IFS= read -r line
 do
-  echo "> ${CDIR}/${line}"
-  if [ -d "${CDIR}/${line}" ]; then
-    rm -R "${CDIR}/${line}"
+  if [ "x${line}" != "x" ]; then
+    echo "> ${CDIR}/${line}"
+    if [ -d "${CDIR}/${line}" ]; then
+      rm -R "${CDIR}/${line}"
+    fi
   fi
 done < ".gitignore"
