+++ update.sh
@@ -4,8 +4,10 @@ echo "!!! cleaning"
 
 IFS=$'\n'; for i in $(git ls-files --others --exclude-standard)
 do
-    echo "!!! delete ${i}"
-    rm -f "${i}"
+    if [ "x${i}" != "x" ]; then
+        echo ">${i}"
+        rm -f "${i}"
+    fi
 done
 
 cp -u .xinitrc ${HOME}/
